Sick LD methods are implemented but not tested

+++ generic/LaserScannerData.cpp
@@ -23,7 +23,7 @@ quantity<length> LaserScannerData::getRanges() const {
   // Bouml preserved body end 0001F77C
 }
 
-unsigned int* LaserScannerData::getRangesPointer() {
+double* LaserScannerData::getRangesPointer() {
   // Bouml preserved body begin 000249F1
   return this->ranges;
   // Bouml preserved body end 000249F1
@@ -39,7 +39,7 @@ void LaserScannerData::setNumMeasurementValues(const unsigned int& num_range_mea
   if(this->ranges != NULL){
     delete this->ranges;
   }
-  this->ranges = new unsigned int[num_range_measurements_values];
+  this->ranges = new double[num_range_measurements_values];
   this->num_range_measurements = num_range_measurements_values;
   // Bouml preserved body end 00024A71
 }
